Module mail_headers::data

A number of little helper types, which contain text.

They provide mainly following functionality:

remember if the data is Ascii/Utf8

this might be extended at some point to contain non ascii data

make sure the types are cheap to clone, by sharing the text internally.

this is mainly helpful when parsing a mail

Both main points are for features which I decided to to not yet implement, as such there is a chance that this module will be removed int the future.

Structs

Input
a Input is similar to Item a container data container used in different context's with different restrictions, but different to an Item it might contain characters which require encoding (e.g. encoded words) to represent them

Enums

InnerAscii
InnerAscii is string data container which can contain either a owned SoftAsciiString or a SoftAsciiStr reference into a shared string buffer.

InnerUtf8
InnerUtf8 is string data container which can contain either a owned String or a str reference into a shared string buffer.

SimpleItem