Struct mail_headers::data::Input

pub struct Input(pub InnerUtf8);

a Input is similar to Item a container data container used in different context's with different restrictions, but different to an Item it might contain characters which require encoding (e.g. encoded words) to represent them

Methods

impl Input

pub fn into_shared(self) -> Self

pub fn into_ascii_item(self) -> StdResult<InnerAscii, Input>

pub fn into_ascii_item_unchecked(self) -> InnerAscii

pub fn into_utf8_item(self) -> InnerUtf8