Module mail_headers::header_components

This modules contains all components provided by this library.

A mail (header) component is basically the body of a header field in the mails header section. E.g. in Subject: Hy There, the Hy There would be represented by an component (in this case the Unstructured component).

Re-exports

pub use self::word::Word;

Modules

utils

word

Structs

DateTime
A DateTime header component wrapping chrono::DateTimechrono::Utc

Disposition
Disposition Component mainly used for the Content-Disposition header (rfc2183)

Domain

Email
an email of the form local-part@domain corresponds to RFC5322 addr-spec, so <, > padding is not part of this Email type (but of the Mailbox type instead)

FWS

FileMeta
A struct representing common file metadata.

LocalPart

Mailbox

MailboxList

MediaType

MessageId
Implementation Details

MessageIdList

NoDisplayName

OptMailboxList

Path

Phrase
Represent a "phrase" as it for example is used in the Mailbox type for the display name.

PhraseList

RawUnstructured
A unstructured header field implementation which validates the given input but does not encode any utf8 even if it would have been necessary (it will error in that case) nor does it support breaking longer lines in multiple ones (no FWS marked for the encoder)

Unstructured

Enums

CFWS

DispositionKind
Represents what kind of disposition is used (Inline/Attachment)

ReceivedToken

TransferEncoding
The TransferEnecoding header component mainly used by the ContentTransferEncodingHeader.

Type Definitions

ContentId

ContentIdList

ContentType