Module mail_headers::headers

Structs

Bcc
(rfc5322)

Cc
(rfc5322)

Comments
(rfc5322)

ContentDescription
A description of the content of the body (rfc2045)

ContentDisposition
Defines the disposition of a multipart part it is used on (rfc2183)

ContentId
(rfc2045)

ContentTransferEncoding
The transfer encoding used to (transfer) encode the body (rfc2045)

ContentType
(rfc2045)

Date
(rfc5322)

InReplyTo
(rfc5322)

Keywords
(rfc5322)

MessageId
(rfc5322)

Received
(rfc5322)

References
(rfc5322)

ReplyTo
(rfc5322)

ResentBcc
(rfc5322)

ResentCc
(rfc5322)

ResentDate
(rfc5322)

ResentFrom
(rfc5322)

ResentMsgId
(rfc5322)

ResentSender
(rfc5322)

ResentTo
(rfc5322)

ReturnPath
(rfc5322)

Sender
(rfc5322)

Subject
(rfc5322)

_From
(rfc5322)

_To
(rfc5322)