Module mail_headers::map

Module containing the HeaderMap.

It also contains some helper types like iterator types for the HeaderMap etc.

Structs

HeaderMap
A header map is a collection representing a number of mail headers in an specific order.

TypedBodies
Iterator over all boxed bodies for a given header name with knows which type they should have

TypedBodiesMut
Iterator over all boxed bodies for a given header name with knows which type they should have

Functions

check_header_count_max_one

Type Definitions

HeaderMapValidator
The type of an validator used to check more complex header contraints.

IntoIter

Iter

IterMut

UntypedBodies
Iterator over all boxed bodies for a given header name

UntypedBodiesMut

Values

ValuesMut